backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pillow
ultralytics
openai
